+++ src/repositories/index.ts
@@ -1,4 +1,4 @@
-import { v4 as uuidv4 } from 'uuid';
+import { randomUUID as uuidv4 } from 'node:crypto';
 import { UserData, UsersDB } from '@/models';
 import { STATUS, STATUS_MESSAGES } from '@/constants';
 import { Repository } from '@/types';
