+++ src/router/routes.ts
@@ -1,10 +1,11 @@
-import { DYNAMIC_PATH } from '@/constants';
 import { ControllerProps } from '@/types';
 
 export interface RouteNode {
 	children: Record<string, RouteNode>;
 	handler?: (args: ControllerProps) => Promise<void>;
 }
+
+export const DYNAMIC_PATH = ':id';
 
 export const ROUTES = {
 	API: {
@@ -13,6 +14,6 @@ export const ROUTES = {
 			ID: `api/users/${DYNAMIC_PATH}`,
 		},
 	},
-};
+} as const;
 
 export const routes: RouteNode = { children: {} };
