+++ src/server.ts
@@ -5,7 +5,6 @@ import { DB } from './db';
 import { UserRepository } from '@/repositories';
 import { UserService } from '@/services';
 import { UsersController } from '@/controllers';
-import { logger } from './utils/logger';
 import { ROUTES } from './router/routes';
 import { argv } from 'node:process';
 import { ClusterUserRepository } from './repositories/cluster-user-repository';
@@ -17,6 +16,7 @@ const usersController = new UsersController(usersService);
 
 export const server = http.createServer(async (req, res) => {
 	const { url, method } = req;
+
 	if (!url || !method || !(method in HTTP_METHODS)) {
 		res.writeHead(STATUS.BAD_REQUEST, { 'Content-Type': 'text/plain' });
 		res.end(STATUS_MESSAGES[STATUS.BAD_REQUEST].badRequest);
@@ -24,26 +24,28 @@ export const server = http.createServer(async (req, res) => {
 		try {
 			const handler = findHandler({ url, method: method as HttpMethods });
 			await handler({ req, res });
-			logger(`[${method}] ${url} status: ${res.statusCode}`);
+			console.log(`[${method}] ${url} status: ${res.statusCode}`);
 		} catch (error) {
 			if (error instanceof Error) {
 				const [statusCode, message] = error.message.split('||');
 				const status = Number.isNaN(Number(statusCode)) ? STATUS.INTERNAL_SERVER_ERROR : Number(statusCode);
-				res.writeHead(Number(status), { 'Content-Type': 'text/plain' });
-				res.end(
-					status === STATUS.INTERNAL_SERVER_ERROR ? STATUS_MESSAGES[STATUS.INTERNAL_SERVER_ERROR] : message
-				);
-				logger(`[${method}]${url} status: ${statusCode}`);
+				res.writeHead(Number(status), { 'Content-Type': 'application/json' });
+
+				const messageBody =
+					status === STATUS.INTERNAL_SERVER_ERROR ? STATUS_MESSAGES[STATUS.INTERNAL_SERVER_ERROR] : message;
+
+				res.end(JSON.stringify({ message: messageBody }));
+				console.log(`[${method}]${url} status: ${statusCode}`);
 			}
 		}
 	}
 });
 
-addRoute('GET', ROUTES.API.USERS.ROUTE, usersController.getUsers);
-addRoute('GET', ROUTES.API.USERS.ID, usersController.getById);
-addRoute('POST', ROUTES.API.USERS.ROUTE, usersController.create);
-addRoute('PUT', ROUTES.API.USERS.ID, usersController.update);
-addRoute('DELETE', ROUTES.API.USERS.ID, usersController.delete);
+addRoute(HTTP_METHODS.GET, ROUTES.API.USERS.ROUTE, usersController.getUsers);
+addRoute(HTTP_METHODS.GET, ROUTES.API.USERS.ID, usersController.getById);
+addRoute(HTTP_METHODS.POST, ROUTES.API.USERS.ROUTE, usersController.create);
+addRoute(HTTP_METHODS.PUT, ROUTES.API.USERS.ID, usersController.update);
+addRoute(HTTP_METHODS.DELETE, ROUTES.API.USERS.ID, usersController.delete);
 
 export function startServer(port: number) {
 	server.listen(port, () => {
