+++ src/single-server.ts
@@ -3,6 +3,6 @@ import { startServer } from './server';
 
 dotenv.config();
 
-export const PORT = Number(process.env.PORT ?? 3000);
+export const PORT = Number(process.env.PORT ?? 4000);
 
 startServer(PORT);
