--- src/utils/logger.ts
@@ -1,3 +1,0 @@
-export function logger(message: string) {
-	console.log(message);
-}
